Prevent default event and propagation on steppers
This kills two birds with one stone. `e.preventDefault()` keeps `#null`
from being appended to the URL, of course, but the more important event
method here is `stopPropagation`:
https://developer.mozilla.org/en-US/docs/Web/API/Event/stopPropagation

The reason that the steppers *appeared* to be giving focus to the input
has nothing to do with the native behavior of `<input>` elements that
are changed programmatically (that is, through JavaScript). What was at
fault here was the the listener on `.inputs li`, which sets the focus on
the `<input>` element if someone clicks anywhere within the `<li>`
element. That includes, in this case, the little stepper buttons.

So, by calling `e.stopPropagation()`, the click event does not bubble
any higher up the DOM tree, to the problem `<li>`. That means that
keyboard users can continue to hit Return to increase or decrease the
input value with the stepper buttons.