Loop through the <options> and build a list
This uses the jQuery .each() function to loop through each item in a
collection. In this case, `$('#size option')` grabs all of the
`<option>` elements inside of the `#size` `<select>` element. Each list
item is then appended to the `<ul class="options">` list, which
preserves the order of options as specified in the list of `<option>`s.