Use event delegation to keep the AJAX party going
From the jQuery docs:

  > Event handlers are bound only to the currently selected elements;
  > they must exist at the time your code makes the call to .on().

So any elements that get loaded later, say using AJAX and the .load()
method, are not going to have the event attached. To do that, we have to
use event delegation. Again, from the jQuery docs:

  > Delegated events have the advantage that they can process events
  > from descendant elements that are added to the document at a later
  > time.

The .on() method takes an optional `selector` argument. One more time
from the jQuery docs:

  > When a selector is provided, the event handler is referred to as
  > delegated. The handler is not called when the event occurs directly
  > on the bound element, but only for descendants (inner elements) that
  > match the selector. jQuery bubbles the event from the event target
  > up to the element where the handler is attached (i.e., innermost to
  > outermost element) and runs the handler for any elements along that
  > path matching the selector.

So rather than binding the 'click' event on `.button` itself, i.e.,

  $('.button').on('click', function() { /* whatever */ });

We delegate, using the `#primary` element and passing in the `.button`
as the selector:

  $('#primary').on('click', '.button', function() { /* whatever */ });

All the quoted material above is from the jQuery docs for .on():
http://api.jquery.com/on/